Replace Algolia search with custom flexsearch implementation
Fetching of the search file and the indexing/search logic happens in the background, on a SharedWorker to avoid blocking the main browser thread and allowing the page to load before the whole search file is fetched.
The SharedWorker is instantiated alongside the store because it makes it easy to manage the worker lifecycle.